requirements.txt
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
black==22.12.0
bleach==6.3.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coverage==7.13.4
crispy-bootstrap5==0.7
cryptography==46.0.2
cssselect2==0.8.0
Deprecated==1.2.18
Django==5.2.8
django-browser-reload==1.21.0
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-crispy-forms==1.14.0
django-elasticsearch-dsl==8.2
django-environ==0.13.0
django-extensions==3.2.3
django-filter==25.2
django-jet-reboot==1.3.5
django-js-asset==3.1.2
django-model-utils==4.3.1
django-modeltranslation==0.18.11
django-mptt==0.18.0
django-oauth-toolkit==3.2.0
django-quill-editor==0.1.42
django-redis==6.0.0
django-summernote==0.8.20.0
django-tailwind==4.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-extensions==0.8.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.3.1
elastic-transport==8.17.1
elasticsearch==8.19.3
elasticsearch-dsl==8.18.0
factory_boy==3.3.1
Faker==37.11.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-DebugToolbar==0.13.1
google-auth==2.49.0
gopay==2.0.1
gunicorn==21.2.0
html5lib==1.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
lxml==6.0.2
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
oscrypto==1.3.0
packaging==25.0
pandas==2.1.2
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==1.10.24
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.1.1
pytest==9.0.2
pytest-cov==7.0.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
reportlab==4.0.4
requests==2.31.0
rpds-py==0.29.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
stripe==5.5.0
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==2.3.7
whitenoise==6.2.0
wrapt==1.17.3
xhtml2pdf==0.2.15
